+++ tracer/src/Datadog.Trace/Debugger/LiveDebuggerFactory.cs
@@ -23,7 +23,7 @@ internal class LiveDebuggerFactory
 
     public static LiveDebugger Create(IDiscoveryService discoveryService, IRemoteConfigurationManager remoteConfigurationManager, ImmutableTracerSettings tracerSettings, string serviceName)
     {
-        var settings = DebuggerSettings.FromSource(GlobalSettings.ConfigurationSource);
+        var settings = DebuggerSettings.FromDefaultSource();
         if (!settings.Enabled)
         {
             Log.Information("Live Debugger is disabled. To enable it, please set DD_DEBUGGER_ENABLED environment variable to 'true'.");
